requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile ./requirements.in -o ./requirements.txt
anyio==4.12.0
    # via httpx
beautifulsoup4==4.14.3
    # via -r ./requirements.in
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
greenlet==3.3.0
    # via sqlalchemy
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via python-telegram-bot
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
python-dateutil==2.9.0.post0
    # via -r ./requirements.in
python-dotenv==1.2.1
    # via -r ./requirements.in
python-telegram-bot==22.5
    # via -r ./requirements.in
requests==2.32.5
    # via -r ./requirements.in
six==1.17.0
    # via python-dateutil
soupsieve==2.8.1
    # via beautifulsoup4
sqlalchemy==2.0.45
    # via -r ./requirements.in
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   sqlalchemy
urllib3==2.6.2
    # via requests
